+++ Assets/Scripts/Resources/ResourceDefaultValues.cs
@@ -8,6 +8,6 @@
         public const float Fuel = 3000;
         public const float MaxFuel = 3000;
         public const float Metal = 100;
-        public const float Ammo = 0;
+        public const int Ammo = 0;
     }
 }
